Add Ability for Custom Setting Post Type
Adds the ability to update settings not handled by theme_mod or options. This is a default Word Press customizer feature that's missing from Unyson. 
This way a custom hook can be used when the switch in $WP_Customize_Setting->update() resolves to default and returns:
do_action( 'customize_update_' . $this->type, $value, $this );