introduce fw.options.* API
This lays the ground for the upcoming Reactive Option Types feature.
For now, just a beautiful API for triggering and listening to events for
particular option types.

Comes with support for triggering and listening to change events.
Also provides a way to filter by contexts (more on contexts later),
which is a very powerful mechanism for advanced option types, allowing
you to compose them in very flexible ways.