Add prefix value for "unique"
Useful opportunity to work with these IDs. For example, if a direct transfer of an ID with a number at the beginning to SCSS - the layout may break (due to incorrect recognition of such ID).
It is also useful for direct transmission of ID blocks/sections. Without the need to add prefixes in several places.